refactor: Simplify ContactComponent by removing redundant error handling - Removed unnecessary catchError in ngOnInit. - Leveraged error handling from service methods. - Consolidated subscriptions for easier management and cleanup.

+++ src/app/home/ui/components/contact/contact.component.ts
@@ -1,9 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Contact } from '../../../../shared/interface/contact.interface';
 import { ContactService } from '../../../data-access/contact.service';
-import { catchError, switchMap } from 'rxjs/operators';
+import { switchMap } from 'rxjs/operators';
 import { CommonModule } from '@angular/common';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-contact',
@@ -12,9 +13,10 @@ import { CommonModule } from '@angular/common';
   templateUrl: './contact.component.html',
   styleUrl: './contact.component.css',
 })
-export class ContactComponent implements OnInit {
+export class ContactComponent implements OnInit, OnDestroy {
   userId: string = '';
   contact: Contact | undefined;
+  private subscription = new Subscription();
 
   constructor(
     private route: ActivatedRoute,
@@ -23,51 +25,54 @@ export class ContactComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    this.route.paramMap
-      .pipe(
-        switchMap((params) => {
-          this.userId = params.get('id') || '';
-          this.contactService.selectContact(this.userId)
-          return this.contactService.getContactById(this.userId);
-        }),
-      )
-      .subscribe((contact) => { this.contact = contact });
+    this.subscription.add(
+      this.route.paramMap
+        .pipe(
+          switchMap((params) => {
+            this.userId = params.get('id') || '';
+            this.contactService.selectContact(this.userId)
+            return this.contactService.getContactById(this.userId);
+          })
+        ).subscribe((contact) => { this.contact = contact })
+    );
   }
 
-  handleFavorite() {
+  handleFavorite():void {
     if (this.contact) {
       const newFavoriteStatus = !this.contact.favorite;
-      this.contactService
-        .updateContact(this.userId, { favorite: newFavoriteStatus })
-        .subscribe({
-          next: (updatedContact) => {
-            this.contact = updatedContact;
-            this.contact.favorite = newFavoriteStatus;
-          },
-          error: (err) => {
-            console.error('Error:', err);
-          },
-        });
+      this.subscription.add(
+        this.contactService
+          .updateContact(this.userId, { favorite: newFavoriteStatus })
+          .subscribe({
+            next: (updatedContact) => {
+              this.contact = updatedContact;
+              this.contact.favorite = newFavoriteStatus;
+            },
+            error: (err) => {
+              console.error('Error updating contact:', err);
+            },
+          })
+      )
     }
   }
 
   handleDelete() {
-    if (this.contact) {
-      if (
-        confirm(
-          `Are you sure you want to delete ${this.contact.name} ${this.contact.lastname}?`
-        )
-      ) {
+    if (this.contact && confirm(`Are you sure you want to delete ${this.contact.name} ${this.contact.lastname}?`)) {
+      this.subscription.add(
         this.contactService.deleteContact(this.userId).subscribe({
           next: (contact) => {
             alert(`${contact.name} ${contact.lastname} has been deleted.`);
             this.router.navigate(['']);
           },
           error: (err) => {
-            console.error('Error:', err);
+            console.error('Error deleting contact:', err);
           },
-        });
-      }
+        })
+      )
     }
+  }
+
+  ngOnDestroy(): void {
+    this.subscription.unsubscribe()
   }
 }
